+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!--
 ## Unreleased
 ### Added
+- Use buttons for Whatsapp Templates instead of quick replies
 ### Removed
 ### Fixed
 ### Security